First upload
This is a modified Contrast Adaptive Sharpening widget. It replaces the original CAS and rescales the "Sharpness" slider value in options to between 0-1. Note that the options menu reading will not change, this only rescales the actual calculation in the widget.

The goal of this rescaling is to allow reduction to sharpness that causes a Moiré pattern artifact on certain maps with gray-ish looking background textures such as on DSDR and Comet Catcher. Setting the Sharpness slider value between 0.8-0.1 or lower is usually sufficient, although it may differ for different monitor pixel layouts and different maps.