First upload
This widget gives you a bindable key to copy an existing building or unit's blueprint and set it to active build command (only if at least one of your selected unit is a builder that can build it). Default key set to shift-q.

If you copy on empty ground, the previously successfully copied blueprint will be used again.